Update the auth_cookie when a file is downloaded explicitly.
Nowadays Roxen will expire an authentication sooner or later. When that happens you had to manually delete the file from RAL and then download it again. This fix will update the auth_cookie from the newly downloaded xrl2-file so the step of having to delete the file manually is removed.

We also issue a better error message if you try to re-fetch a file with an expired authentication directly from RAL.